requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
colorama==0.4.6
filelock==3.18.0
fsspec==2025.3.2
huggingface-hub==0.31.2
idna==3.10
jinja2==3.1.6
joblib==1.5.0
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.5
packaging==25.0
pillow==11.2.1
protobuf==6.30.2
pyyaml==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
setuptools==80.4.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
typing-extensions==4.13.2
urllib3==2.4.0
